# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant API built with hexagonal architecture. Provides identity management, tenant management, email templates, and more.",
        "title": "Hexagonal Go API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/email/templates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of email templates with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "List email templates",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Filter by tenant ID",
                        "name": "tenant_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "draft",
                            "active",
                            "archived",
                            "deleted"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "en",
                        "description": "Filter by locale",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by slug containing text",
                        "name": "slug_contains",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name containing text",
                        "name": "name_contains",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "Include system templates",
                        "name": "include_system",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "updated_at",
                            "name",
                            "slug"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of templates",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.ListTemplatesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new email template with subject, body, and variable definitions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Create email template",
                "parameters": [
                    {
                        "description": "Create template request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Template created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.CreateTemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Template with slug already exists",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/email/templates/by-slug": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves an email template by its slug and locale",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Get email template by slug",
                "parameters": [
                    {
                        "type": "string",
                        "example": "welcome-email",
                        "description": "Template slug",
                        "name": "slug",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "en",
                        "example": "en",
                        "description": "Locale code",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tenant ID",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template found",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.GetTemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Slug is required",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/email/templates/preview-by-slug": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Renders an email template by slug with provided data for preview",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Preview email template by slug",
                "parameters": [
                    {
                        "description": "Preview request with slug and data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.PreviewTemplateBySlugRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rendered template preview",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.PreviewTemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing slug",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/email/templates/{id}/activate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Activates a draft or archived template, making it available for use",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Activate email template",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template activated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.ActivateTemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid template ID",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid status transition",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/email/templates/{id}/archive": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Archives an active template, removing it from active use",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "email"
                ],
                "summary": "Archive email template",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template archived successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.ArchiveTemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid template ID",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Template not found",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid status transition",
                        "schema": {
                            "$ref": "#/definitions/internal_email_interface_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_0xsj_hexagonal-go_internal_email_application_dto.CreateTemplateResponse": {
            "type": "object",
            "properties": {
                "template": {
                    "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.TemplateDTO"
                }
            }
        },
...
        "github_com_0xsj_hexagonal-go_internal_email_application_dto.ListTemplatesResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_email_application_dto.TemplateDTO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "internal_email_interface_http_v1.ErrorResponse": {
            "description": "Error response returned by the API",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "TEMPLATE_NOT_FOUND"
                },
                "message": {
                    "type": "string",
                    "example": "template not found"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}